+++ concordium-sdk/src/main/java/com/concordium/sdk/transactions/TransactionFactory.java
@@ -56,4 +56,15 @@ public class TransactionFactory {
     public static UpdateContractTransaction.UpdateContractTransactionBuilder newUpdateContract() {
         return UpdateContractTransaction.builder();
     }
+
+    /**
+     * Creates a new {@link DeployModuleTransaction.DeployModuleTransactionBuilder} for
+     * creating a {@link DeployModuleTransaction}
+     *
+     * @return the builder for a {@link DeployModuleTransaction}
+     */
+    public static DeployModuleTransaction.DeployModuleTransactionBuilder newDeployModule() {
+        return DeployModuleTransaction.builder();
+    }
+
 }