+++ concordium-sdk/src/main/java/com/concordium/sdk/transactions/TransactionType.java
@@ -22,7 +22,10 @@ public enum TransactionType {
     REGISTER_DATA((byte) 21),
     TRANSFER_WITH_MEMO((byte) 22);
 
-    static final int BYTES = 1;
+    /**
+     * Number of Bytes used for Serializing {@link TransactionType}.
+     */
+    public static final int BYTES = 1;
 
     @Getter
     private final byte value;